refactor core - move DiagnosticLogger, _InternalLogMessage, Notification to a wrapper class from core
