+++ shared/AppInsightsCore/src/JavaScriptSDK.Interfaces/INotificationManager.ts
@@ -1,0 +1,37 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+import { ITelemetryItem } from "../JavaScriptSDK.Interfaces/ITelemetryItem";
+import { INotificationListener } from "../JavaScriptSDK.Interfaces/INotificationListener";
+
+/**
+ * Class to manage sending notifications to all the listeners.
+ */
+export interface INotificationManager {
+    listeners: INotificationListener[];
+
+    /**
+     * Adds a notification listener.
+     * @param {INotificationListener} listener - The notification listener to be added.
+     */
+    addNotificationListener(listener: INotificationListener): void;
+
+    /**
+     * Removes all instances of the listener.
+     * @param {INotificationListener} listener - AWTNotificationListener to remove.
+     */
+    removeNotificationListener(listener: INotificationListener): void;
+
+    /**
+     * Notification for events sent.
+     * @param {ITelemetryItem[]} events - The array of events that have been sent.
+     */
+    eventsSent(events: ITelemetryItem[]): void;
+
+    /**
+     * Notification for events being discarded.
+     * @param {ITelemetryItem[]} events - The array of events that have been discarded by the SDK.
+     * @param {number} reason           - The reason for which the SDK discarded the events. The EventsDiscardedReason
+     * constant should be used to check the different values.
+     */
+    eventsDiscarded(events: ITelemetryItem[], reason: number): void;
+}
